# Optimize Enum.to_list for ranges, closes #12383